+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * Set all weights to 0 for CHRONOLOGICAL preset
+* Remove `Toot.reblogsByAccts()` method
 
 ### v0.19.6
 * Remove `Toot.attachmentPrefix()` method